requirements.txt
Django==1.10.3
django-bootstrap3==7.1.0
django-crispy-forms==1.6.1
